### 2025-01-10 #7641 @editorialbot
[PRE REVIEW]: 3DQLayers: Volumetric Layer Based Analysis for Quantitative Renal MRI
**Submitting author:** @alexdaniel654 (<a href="http://orcid.org/0000-0003-2353-3283">Alexander Daniel</a>)
**Repository:** https://github.com/alexdaniel654/qlayers
**Branch with paper.md** (empty if default branch): paper
**Version:** v0.0.3
