### 2025-01-10 #7641 @Kevin-Mattheus-Moerman
- [ ] Figure 2c says `pelvise`, should that read `pelvis`?

In addition, please note that currently managing a large backlog of submissions and the editors most appropriate for your area are all already rather busy.
