### 2025-01-22 920b6c2
Added contributing guide and code of conduct
