### 2025-01-22 62e2636
Fix typo in fig 2 and add DOI
