+++ paper/paper.bib
@@ -270,6 +270,7 @@
   eprint = {at6LBQAAQBAJ},
   eprinttype = {googlebooks},
   publisher = {Springer Science \& Business Media},
+  doi = {10.1007/978-1-4614-3785-7},
   isbn = {978-94-011-4086-7},
   langid = {english},
   pagetotal = {255}
