### 2025-02-07 #7641 @alexdaniel654
Hi @Kevin-Mattheus-Moerman apologies for the delay.

I've added a DOI for Principles of Renal Physiology, couldn't track one down for Guyton and Hall Textbook of Medical Physiology and Trimesh [doesn't have one](https://github.com/mikedh/trimesh/issues/2166). I've also corrected the pelvis typo. These changes are now in the paper branch on the repo.
I've added a [contributing guide](https://github.com/alexdaniel654/qlayers/blob/main/.github/CONTRIBUTING.md) to the main (and paper) branch.

Some suggested reviewers are Eric Sigmund (Department of Radiology, Center for Advanced Imaging Innovation and Research (CAI2R), Center for Biomedical Imaging, NYU Langone Health, New York, USA), Frank Zöllner (Computer Assisted Clinical Medicine, Mannheim Institute for Intelligent Systems in Medicine (MIISM), Medical Faculty Mannheim, Heidelberg University, Germany) and Matthias Stuber (Center for Biomedical Imaging, University Hospital Lausanne, Lausanne, Switzerland).

Thanks!
