### 2025-04-14 #7641 @Kevin-Mattheus-Moerman
@grlee77, @agahkarakuzu, @mfroeling, @agahkarakuzu, @MaierOli2010, @mathieuboudreau, @paquiteau, would you be interested in reviewing this submission for [JOSS (the Journal of Open Source Software](https://joss.theoj.org)? The work is titled `3DQLayers: Volumetric Layer Based Analysis for Quantitative Renal MRI`.

JOSS reviews take place on GitHub and focus on [the software :floppy_disk: ](https://github.com/alexdaniel654/qlayers) as well as [a short paper :scroll: ](https://raw.githubusercontent.com/openjournals/joss-papers/joss.07641/joss.07641/10.21105.joss.07641.pdf). We typically aim for a 2-3 week review but we can be more flexible if needed. You can find more about our review process in our [review guidelines](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html).  Here is an example of an ongoing review if you want to see what it is like: https://github.com/openjournals/joss-reviews/issues/7734.

If you are interested you can let me know by responding here.

Thanks!
