+++ binder/requirements.txt
@@ -7,3 +7,5 @@ scikit-image
 tqdm
 trimesh
 qlayers
+matplotlib
+seaborn
