### 2025-05-09 #8121 @agahkarakuzu
The codebase developed by @alexdaniel654 and colleagues provides a robust, well-structured Python library for quantifying anatomical layers in kidney MRI data. This is a welcome contribution, as most work combining morphometry analysis with quantitative metrics has focused primarily on brain imaging. From this standpoint, the project adds meaningful value to the open-source MRI community.

The codebase is modular, with a clear public API and comprehensive docstrings throughout. The main functionality (layer quantification, integration of quantitative maps, and advanced analyses such as cortical thickness estimation) is implemented using standard scientific Python libraries and is supported by extensive, well-designed tests and code coverage.

All the project metadata files are in place, as well as automated pypi releases. User-facing documentation is intuitive and up-to-date. I was able to download the example datasets and execute all 3 notebooks under the `examples` directory, which generated the same outputs.

To make running tutorials a bit more streamlined, I sent a PR (https://github.com/alexdaniel654/qlayers/pull/8), binderizing code and data requirements for an online executable version. I leave it to the author's discretion to merge it. It would also create an opportunity to slightly tweak the docs (e.g., do `pip install -r binder/requirements.txt` installs all the dependencies to run the example notebooks).

Overall, qlayers is a high-quality, domain-specific library that merits publication in JOSS.

Congrats to the team 🎉
