### 2025-05-22 #8121 @paquiteau
Hi everyone, sorry for the late review, but it is very positive:
- The source code is well written and documented, tested and automated (maybe consider to move to a src-layout and pyproject.toml setup to be more future proof? It's not a blocking point for me)
- The link to the documentation was quite hard to find (its a badge in the readme) I would recommend to add it in the "About" area on the left of the file listing, as well as a dedicated link in the README. Such great docs should be easier to find :)
- The paper conveys the goals of the software clearly.

OVerall, this packages just do the rights things, nothing more, and that's great !
