### 2025-06-14 #8121 @Kevin-Mattheus-Moerman
@paquiteau your `references` box is still unticked. Is there an issue there or can it now be ticked?
