### 2025-06-14 #8121 @paquiteau
> @paquiteau your `references` box is still unticked. Is there an issue there or can it now be ticked?

No issues anymore :)
