+++ setup.py
@@ -10,7 +10,7 @@ with open("README.md", encoding="utf-8") as f:
 
 setup(
     name="qlayers",
-    version="0.0.3",
+    version="1.0.0",
     description="Quantitative layer analysis for renal MRI",
     long_description=long_description,
     long_description_content_type="text/markdown",
