+++ CITATION.cff
@@ -8,6 +8,6 @@ authors:
   given-names: "Susan T"
   orcid: "https://orcid.org/0000-0003-0903-7507"
 title: "QLayers"
-version: 0.0.3
-date-released: 2024-12-02
+version: 1.0.0
+date-released: 2025-06-16
 url: "https://github.com/alexdaniel654/qlayers"
