### 2025-06-27 #8121 @alexdaniel654
@Kevin-Mattheus-Moerman Not 100% sure what's causing that error. My best guess is the unusual date format of one of the references in the `.bib` file that's causing there to be [two `cYear` tags](https://github.com/openjournals/joss-papers/blob/dc504b871d496d06e5a0a04a82190049a0861188/joss.08121/10.21105.joss.08121.crossref.xml#L293). I've [adjusted](https://github.com/alexdaniel654/qlayers/commit/0b1c3272dd92f857c80ecde3f16b424af5202dbd) it in the hope it fixes the issue.
