### 2025-07-11 #8121 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
