### 2025-07-11 #8121 @Kevin-Mattheus-Moerman
@alexdaniel654 congratulations on this JOSS publication!!

I would like to say a special thank you ro the reviewers: @agahkarakuzu, @paquiteau !!!
